Extract PropertyParser from NineChroniclesNodeServiceProperties
