+++ Libplanet.Headless/IceServerInvalidException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace NineChronicles.Headless.Exceptions
+namespace Libplanet.Headless
 {
     [Serializable]
     public class IceServerInvalidException: Exception
